requirements.txt
python-vlc==3
